Break stream and helper classes into submodules
