+++ pyacq/core/stream/__init__.py
@@ -1,2 +1,8 @@
-from .stream import *
+from .stream import InputStream, OutputStream
+from .ringbuffer import RingBuffer
 from .sharedarray import SharedArray
+from .streamhelpers import all_transfermodes, register_transfermode
+
+# import transfer modes so they register their helper classes
+from . import plaindata
+from . import sharedmem
